AI-Doctor-main/AI Medical support by nms/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Pillow==9.0.0
python-dotenv==0.19.0
psycopg2-binary==2.9.1
Werkzeug==2.0.1
gunicorn==20.1.0 